+++ src/main/java/com/simsilica/mathd/Vec3d.java
@@ -118,6 +118,13 @@ public class Vec3d implements Cloneable, java.io.Serializable {
         this.z = v.z;
         return this;
     }
+
+    public final Vec3d set( Vector3f v ) {
+        this.x = v.x;
+        this.y = v.y;
+        this.z = v.z;
+        return this;
+    }
  
     @Override
     public final Vec3d clone() {