+++ release-notes.md
@@ -1,6 +1,7 @@
 Version 1.1.2 (unreleased)
 --------------
 * Added Vec4i.toVec3d()
+* Added Vec3d.set(Vector3f)
 
 Version 1.1.1 
 --------------